[VAS-947] chore: Refactor stations unit tests
