--- src/pages/stations/__tests__/StationsPage.test.tsx
@@ -1,35 +1,0 @@
-import {ThemeProvider} from '@mui/system';
-import {theme} from '@pagopa/mui-italia';
-import {cleanup, render, screen} from '@testing-library/react';
-import {Router} from 'react-router-dom';
-import {store} from '../../../redux/store';
-import {createMemoryHistory} from 'history';
-import {Provider} from 'react-redux';
-import StationPage from '../list/StationsPage';
-
-beforeEach(() => {
-  jest.spyOn(console, 'error').mockImplementation(() => {});
-  jest.spyOn(console, 'warn').mockImplementation(() => {});
-});
-
-afterEach(cleanup);
-
-describe('<StationPage />', () => {
-  const history = createMemoryHistory();
-
-  test('render component StationPage', () => {
-    history.location.state = { alertSuccessMessage: 'Success!' };
-
-    render(
-      <Provider store={store}>
-        <ThemeProvider theme={theme}>
-          <Router history={history}>
-            <StationPage />
-          </Router>
-        </ThemeProvider>
-      </Provider>
-    );
-
-    expect(screen.getByTestId('alert-test')).toBeInTheDocument();
-  });
-});
