+++ src/pages/stations/addEditStation/__tests__/AddEditStationFormSectionTitle.test.tsx
@@ -3,8 +3,8 @@ import {theme} from '@pagopa/mui-italia';
 import {cleanup, render} from '@testing-library/react';
 import React from 'react';
 import {Router} from 'react-router-dom';
-import {store} from '../../../redux/store';
-import AddEditStationFormSectionTitle from '../addEditStation/AddEditStationFormSectionTitle';
+import {store} from '../../../../redux/store';
+import AddEditStationFormSectionTitle from '../AddEditStationFormSectionTitle';
 import {createMemoryHistory} from 'history';
 import {Provider} from 'react-redux';
 import {MenuBook} from '@mui/icons-material';
