+++ src/pages/stations/addEditStation/__tests__/AddEditStationPage.test.tsx
@@ -2,10 +2,11 @@ import { ThemeProvider } from '@mui/system';
 import { theme } from '@pagopa/mui-italia';
 import { cleanup, render } from '@testing-library/react';
 import { Router } from 'react-router-dom';
-import { store } from '../../../redux/store';
+import { store } from '../../../../redux/store';
 import { createMemoryHistory } from 'history';
 import { Provider } from 'react-redux';
-import AddEditStationPage from '../addEditStation/AddEditStationPage';
+import React from 'react';
+import AddEditStationPage from '../AddEditStationPage';
 
 beforeEach(() => {
   jest.spyOn(console, 'error').mockImplementation(() => {});
