+++ src/pages/stations/list/__tests__/StationsTable.test.tsx
@@ -1,12 +1,19 @@
-import {ThemeProvider} from '@mui/system';
-import {theme} from '@pagopa/mui-italia';
-import {cleanup, render} from '@testing-library/react';
+import { ThemeProvider } from '@mui/system';
+import { theme } from '@pagopa/mui-italia';
+import { cleanup, render, waitFor, screen } from '@testing-library/react';
 import React from 'react';
-import {Router} from 'react-router-dom';
-import {store} from '../../../../redux/store';
+import { Router } from 'react-router-dom';
+import { createMemoryHistory } from 'history';
+import { Provider } from 'react-redux';
+import { useAppDispatch } from '../../../../redux/hooks';
+import { partiesActions } from '../../../../redux/slices/partiesSlice';
+import { store } from '../../../../redux/store';
+import * as StationService from '../../../../services/stationService';
+import { pspAdminSignedDirect } from '../../../../services/__mocks__/partyService';
+import { mockedStationsMerged2 } from '../../../../services/__mocks__/stationService';
 import StationsTable from '../StationsTable';
-import {createMemoryHistory} from 'history';
-import {Provider} from 'react-redux';
+
+const mockGetStationsMerged = jest.spyOn(StationService, 'getStationsMerged');
 
 beforeEach(() => {
   jest.spyOn(console, 'error').mockImplementation(() => {});
@@ -15,18 +22,43 @@ beforeEach(() => {
 
 afterEach(cleanup);
 
-describe('<StationsTable />', () => {
+const Component = () => {
   const history = createMemoryHistory();
+  const dispatch = useAppDispatch();
+  dispatch(partiesActions.setPartySelected(pspAdminSignedDirect));
 
-  test('render component StationsTable', () => {
+  return (
+    <ThemeProvider theme={theme}>
+      <Router history={history}>
+        <StationsTable stationCode={'1'} />
+      </Router>
+    </ThemeProvider>
+  );
+};
+
+describe('<StationTable />', () => {
+  test('render component StationTable', async () => {
+    mockGetStationsMerged.mockReturnValueOnce(Promise.resolve(mockedStationsMerged2));
     render(
       <Provider store={store}>
-        <ThemeProvider theme={theme}>
-          <Router history={history}>
-            <StationsTable  stationCode={"1"}/>
-          </Router>
-        </ThemeProvider>
+        <Component />
       </Provider>
     );
+
+    await waitFor(() => {
+      expect(screen.queryByTestId('empty-state-table')).not.toBeInTheDocument();
+    });
+  });
+  test('render component StationTable error getStationMerged', async () => {
+    mockGetStationsMerged.mockRejectedValueOnce('error');
+    render(
+      <Provider store={store}>
+        <Component />
+      </Provider>
+    );
+
+    await waitFor(() => {
+      expect(screen.queryByTestId('empty-state-table')).toBeInTheDocument();
+    });
   });
 });
